Some final touches and fixes for filter search
- Renamed SearchResults component to SearchResultsOverview
- Renamed SearchResultsComponent component to SearchResults
- Removed disabled state from SelectFilter
- Fixed disabled state for RadioFilter and CheckboxFilter
- Fixed bug where 'Show more' and 'Show all' buttons were swapped on the data search page
- Changed RadioFilter and CheckboxFilter to use a primary color variant
- Result counts are now shown in filters on the data search page
- Changed data filter type to match the one on the API (from 'types' to 'dataTypes')
