+++ src/app/pages/SearchPage/SearchPageFilters.tsx
@@ -5,13 +5,23 @@ import { Filter } from '../../models'
 interface SearchPageFiltersProps {
   filters: Filter[]
   totalCount: number
+  hideCount: boolean
 }
 
-const SearchPageFilters: React.FC<SearchPageFiltersProps> = ({ filters, totalCount }) => {
+const SearchPageFilters: React.FC<SearchPageFiltersProps> = ({
+  filters,
+  totalCount,
+  hideCount,
+}) => {
   return (
     <>
       {filters.map(filter => (
-        <SearchFilter key={filter.type} filter={filter} totalCount={totalCount} hideCount />
+        <SearchFilter
+          key={filter.type}
+          filter={filter}
+          totalCount={totalCount}
+          hideCount={hideCount}
+        />
       ))}
     </>
   )
