+++ src/app/pages/SearchPage/SearchPageResults.jsx
@@ -13,8 +13,8 @@ import { Enlarge } from '@datapunt/asc-assets'
 import PAGES from '../../pages'
 import SEARCH_PAGE_CONFIG, { EDITORIAL_SEARCH_PAGES } from './config'
 import ActionButton from '../../components/ActionButton/ActionButton'
-import SearchResults, { SearchResultsSkeleton } from './SearchResultsComponent'
-import SearchResultsOverview, { SearchResultsOverviewSkeleton } from './SearchResults'
+import SearchResults, { SearchResultsSkeleton } from './SearchResults'
+import SearchResultsOverview, { SearchResultsOverviewSkeleton } from './SearchResultsOverview'
 import SearchSort from './SearchSort'
 import { DEFAULT_LOCALE } from '../../../shared/config/locale.config'
